feat: runtime-lookup-factory-config-capability-cleanup-002 - Migrate topology projection and verify factory name behavior
